refactor: modify entity file name - 엔티티 파일명을 클래스명과 동일하게 변경
